+++ src/NTwitch.Core/Entities/Channels/ISelfChannel.cs
@@ -9,7 +9,7 @@ namespace NTwitch
         /// <summary> Create a new post on this channel. </summary>
         Task CreatePostAsync();
         /// <summary> Delete an existing post on this channel. </summary>
-        Task DeletePostAsync(ulong postId);
+        Task DeletePostAsync(uint postId);
         /// <summary> Reset the stream key on this channel. </summary>
         Task ResetStreamKeyAsync();
         /// <summary> Start a commercial on this channel. </summary>
@@ -19,6 +19,6 @@ namespace NTwitch
         /// <summary> Get a collection of users who are subscribed to this channel. </summary>
         Task GetSubscribersAsync();
         /// <summary> Get a specified subscriber, if it exists. </summary>
-        Task GetSubscriberAsync(ulong userId);
+        Task GetSubscriberAsync(uint userId);
     }
 }
